purr_on_aws/requirements.txt
aws-cdk-lib==2.178.0
boto3
python-dotenv
constructs>=10.0.0,<11.0.0
